tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation_shorthand_infinite_name.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation shorthand, 'infinite' before name</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-shorthand-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="'infinite move 0.2s alternate' is a valid value
   for the animation shorthand">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         from {
            left: 0px;
         }
         to {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: infinite move 0.2s alternate;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1000);
         var iterations = 0;
         div.addEventListener('animationiteration', function() {
            iterations++;
            if (iterations > 2) {
               clearTimeout(timeout);
               test(function(){assert_true(true);}, testName); done();
               log.data = 'PASS if the square animates.';
            }
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should slide from left to right to left,
   indefinitely.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
